chore(Datagrid): add sticky actions example, fix docs page
