+++ packages/cloud-cognitive/src/components/Datagrid/Datagrid.mdx
@@ -22,7 +22,7 @@ import { Datagrid } from '.';
 <!-- TODO: One example per designed use case. -->
 
 <Canvas>
-  <Story id={getStoryId(Datagrid.displayName, 'datagrid')} />
+  <Story id={getStoryId(Datagrid.displayName, 'basic-usage')} />
 </Canvas>
 
 ## Code sample
