+++ packages/cloud-cognitive/src/components/Datagrid/utils/makeData.js
@@ -1,0 +1,79 @@
+/**
+ * Copyright IBM Corp. 2022, 2022
+ *
+ * This source code is licensed under the Apache-2.0 license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+
+import React from 'react';
+import namor from 'namor';
+
+export const makeData = (...lens) => {
+  const makeDataLevel = (depth = 0) => {
+    const len = lens[depth];
+    return range(len).map(() => ({
+      ...newPerson(),
+      subRows: lens[depth + 1] ? makeDataLevel(depth + 1) : undefined,
+    }));
+  };
+
+  return makeDataLevel();
+};
+
+export const range = (len) => {
+  const arr = [];
+  for (let i = 0; i < len; i++) {
+    arr.push(i);
+  }
+  return arr;
+};
+
+const newPerson = () => {
+  const statusChance = Math.random();
+  return {
+    firstName: namor.generate({ words: 1, numbers: 0 }),
+    lastName: namor.generate({ words: 1, numbers: 0 }),
+    age: Math.floor(Math.random() * 30),
+    visits: Math.floor(Math.random() * 100),
+    progress: Math.floor(Math.random() * 100),
+    someone1: namor.generate({ words: 1, numbers: 0 }),
+    someone2: namor.generate({ words: 1, numbers: 0 }),
+    someone3: namor.generate({ words: 1, numbers: 0 }),
+    someone4: namor.generate({ words: 1, numbers: 0 }),
+    someone5: namor.generate({ words: 1, numbers: 0 }),
+    someone6: namor.generate({ words: 1, numbers: 0 }),
+    someone7: namor.generate({ words: 1, numbers: 0 }),
+    someone8: namor.generate({ words: 1, numbers: 0 }),
+    someone9: namor.generate({ words: 1, numbers: 0 }),
+    someone10: namor.generate({ words: 1, numbers: 0 }),
+    someone11: namor.generate({ words: 1, numbers: 0 }),
+    someone12: namor.generate({ words: 1, numbers: 0 }),
+    someone13: namor.generate({ words: 1, numbers: 0 }),
+    someone14: namor.generate({ words: 1, numbers: 0 }),
+    someone15: namor.generate({ words: 1, numbers: 0 }),
+    someone16: namor.generate({ words: 1, numbers: 0 }),
+    someone17: namor.generate({ words: 1, numbers: 0 }),
+    someone18: namor.generate({ words: 1, numbers: 0 }),
+    someone19: namor.generate({ words: 1, numbers: 0 }),
+    someone20: namor.generate({ words: 1, numbers: 0 }),
+    status:
+      statusChance > 0.66
+        ? 'relationship'
+        : statusChance > 0.33
+        ? 'complicated'
+        : 'single',
+  };
+};
+
+export const newPersonWithTwoLines = () => {
+  return {
+    firstName: (
+      <>
+        <div>{namor.generate({ words: 1, numbers: 0 })}</div>
+        <div>{namor.generate({ words: 1, numbers: 0 })}</div>
+      </>
+    ),
+    lastName: namor.generate({ words: 1, numbers: 0 }),
+    age: Math.floor(Math.random() * 30),
+  };
+};
